requirements.txt
tensorflow==2.15.0
numpy==1.24.3
matplotlib==3.7.1
opencv-python==4.8.0.74
Pillow==10.0.0
